% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charNTS}
\alias{charNTS}
\title{Title}
\usage{
charNTS(
  t,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  lambda = NULL,
  mu = NULL,
  theta = NULL
)
}
\arguments{
\item{t}{A gap holder.}

\item{alpha}{A real number between 0 and 1.}

\item{beta}{A gap holder}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
charNTS(0.1,0.9,10,20,30,40)
charNTS(0.1,0.9,1,2,3,4, NULL)

}
