% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{charSTS}
\alias{charSTS}
\title{Function title.}
\usage{
charSTS(t, alpha = NULL, delta = NULL, lambda = NULL, theta = NULL)
}
\arguments{
\item{t}{A positive integer.}

\item{alpha}{A real number between 0 and 1.}

\item{delta}{A real number > 0.}

\item{lambda}{A  real number > 0.}

\item{theta}{A vector of all other arguments.}
}
\value{
The result of spectral positive tempered stable process
}
\description{
When \code{alpha} is a real number between 0 and 1, the tempered stable
distribution TS'(\code{alpha}, \code{delta}, \code{lambda}) can be simulated
exactly through acceptance-rejection sampling.
}
\details{
For whole derivation and meaning of single letters see also Kawai et. Masuda
(2011).
Basically, the derivation starts from a one-dimensional stable distribution
which is later added with a centered and totally positively skewed tempered
stable (Levy) process.

Reiichiro Kawai, & Hiroki Masuda (2011).
On simulation of tempered stable random variates.
Journal of Computational and Applied Mathematics, 235(8), 2873-2887.
https://doi.org/10.1016/j.cam.2010.12.014
}
\examples{
charSTS(1000,0.5,1,0.3)
charSTS(500,0.9,1,0.3)

}
\seealso{
(\url{https://www.sciencedirect.com/science/article/pii/S0377042710006643})
}
