% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{chartocdf}
\alias{chartocdf}
\title{Function title}
\usage{
chartocdf(q, N, m, s, char, ...)
}
\arguments{
\item{q}{A gap holder.}

\item{N}{A gap holder.}

\item{m}{A gap holder.}

\item{s}{A gap holder.}

\item{char}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
chartocdf(0.5,10,1,1,charNTS, alpha=0.5, beta = 1, delta = 1, lambda = 1,
 mu = 1)
chartocdf(0.5,10,1,1,charCTS, alpha=0.5, deltap = 1, deltam = 1, lambdap = 1,
 lambdam = 1, mu = 1)

}
