% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{ComputeStatOutput}
\alias{ComputeStatOutput}
\title{Function title}
\usage{
ComputeStatOutput(
  EstimOutput,
  FctsToApply,
  SampleSizes,
  CheckMat,
  tolFailCheck,
  MCparam,
  ...
)
}
\arguments{
\item{EstimOutput}{A gap holder.}

\item{FctsToApply}{A gap holder.}

\item{SampleSizes}{A gap holder.}

\item{CheckMat}{A gap holder.}

\item{tolFailCheck}{A gap holder.}

\item{MCparam}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
