% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{getTempEstimation}
\alias{getTempEstimation}
\title{No export.}
\usage{
getTempEstimation(
  thetaT,
  x,
  seed,
  size,
  Ncol,
  TemperedType,
  Estimfct,
  HandleError,
  eps,
  ...
)
}
\description{
No export.
}
