% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MasterEstim.R
\name{getTempEstimFcts}
\alias{getTempEstimFcts}
\title{Function title}
\usage{
getTempEstimFcts(
  type = c("Classic", "Subordinator", "Normal", "CGMY"),
  method = c("ML", "GMM", "Cgmm", "GMC")
)
}
\arguments{
\item{type}{A String. Either "Classic", "Subordinator", "Normal", or
"CGMY".}

\item{method}{A String. Either "ML", "GMM", "Cgmm", or "GMC".}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
