% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{pCTS}
\alias{pCTS}
\title{Function title}
\usage{
pCTS(
  q,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  a = -40,
  b = 40,
  nf = 2^13,
  ...
)
}
\arguments{
\item{q}{A gap holder.}

\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdap}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{a}{A gap holder. -20 by default.}

\item{b}{A gap holder. 20 by default.}

\item{nf}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
\examples{
pCTS(0.5,0.5,1,1,1,1,1)
pCTS(0.5,0.9,1,2,3,4,5)

}
