% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicfcts_TS.R
\name{qCTS}
\alias{qCTS}
\title{Function title}
\usage{
qCTS(
  p,
  alpha = NULL,
  deltap = NULL,
  deltam = NULL,
  lambdap = NULL,
  lambdam = NULL,
  mu = NULL,
  theta = NULL,
  ...
)
}
\arguments{
\item{alpha}{A real number between 0 and 2.}

\item{deltap, deltam}{A real number > 0.}

\item{lambdap, lambdap}{A  real number > 0.}

\item{mu}{A location parameter, any real number.}

\item{theta}{A vector of all other arguments.}

\item{q}{A gap holder.}
}
\value{
Gap holder for return.

TODO: Examples
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
