% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MonteCarloFunction.R
\name{readCheckPoint}
\alias{readCheckPoint}
\title{Function title}
\usage{
readCheckPoint(
  ParameterMatrix,
  TemperedType,
  Estimfct,
  nab,
  npar,
  nSS,
  MCparam,
  ...
)
}
\arguments{
\item{ParameterMatrix}{A gap holder.}

\item{TemperedType}{A gap holder.}

\item{Estimfct}{A gap holder.}

\item{nab}{A gap holder.}

\item{npar}{A gap holder.}

\item{nSS}{A gap holder.}

\item{MCparam}{A gap holder.}
}
\value{
Gap holder for return.
}
\description{
Gap holder for description.
}
\details{
Gap holder for details.
}
