% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tillsfunctionpackage.R
\docType{package}
\name{tillsfunctionpackage}
\alias{tillsfunctionpackage}
\title{tillsfunctionpackage: Describe package in one sentence}
\description{
Here is more space to go into detail.
Users can enter this documentation with ?tillsfunctionpackage.
}
\section{Basic Functions}{

Here you can use sections to divide content.
}

